+++ src/types.ts
@@ -29,6 +29,9 @@ export interface SelectOption {
   label: string;
   value: string;
   description?: string;
+  sortIndex?: number;
+  node?: NodeType;
+  internalType?: "comment";
 }
 
 export interface SelectControl extends Control {
